pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FracDimPy"
version = "0.1.3"
authors = [
    {name = "Zhile Han", email = "2667032759@qq.com"},
]
description = "A comprehensive Python package for fractal dimension calculation and multifractal analysis"
readme = "README_EN.md"
requires-python = ">=3.8"
keywords = ["fractal", "fractal dimension", "multifractal", "box-counting", "Hurst", "R/S analysis"]
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "pytest-xdist>=3.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
    "isort>=5.10",
    "bandit>=1.7",
    "safety>=2.0",
    "scienceplots>=2.0.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autodoc-typehints>=1.19",
    "myst-parser>=0.18",
]
all = [
    "FracDimPy[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/Kecoya/FracDimPy"
Documentation = "https://github.com/Kecoya/FracDimPy/blob/main/README.md"
Repository = "https://github.com/Kecoya/FracDimPy"
Issues = "https://github.com/Kecoya/FracDimPy/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fracDimPy = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

